backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
openai==1.3.8
starlette==0.27.0
